updated files and requirements
