+++ requirements.txt
@@ -1,0 +1,3 @@
+pandas
+pymongo
+json
