+++ scripts/info.md
@@ -1,1 +1,1 @@
-
+mongodb pw: Hh2fTPMAJUzzPqiC
